+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add keyboard shortcuts for popup and side panel - [@tunglt1810]
 
 ### Changed
 - enhance Chrome Web Store upload process with detailed logging and status checks - [@tunglt1810]